controllers/requirements.txt
bert-extractive-summarizer==0.7.1
Click==7.1.1
datasets>=1.1.3
nltk
protobuf
rouge-score
sentencepiece==0.1.95
torch==1.8.1
transformers
